+++ src/lib/index.ts
@@ -1,12 +1,30 @@
-export * from './client'
+export {
+    Client,
+    install,
+    fetchScript,
+    queryLoadingGraph,
+    installLoadingGraph,
+} from './client'
 export * from './state'
 export * from './loader.view'
-export * from './models'
 export {
-    getAssetId,
-    getUrlBase,
-    parseResourceId,
-    sanitizeModules,
-} from './utils'
+    FetchScriptInputs,
+    InstallInputs,
+    InstallLoadingGraphInputs,
+    QueryLoadingGraphInputs,
+    LightLibraryQueryString,
+    ModuleSideEffectCallback,
+    ScriptInput,
+    CssInput,
+    CustomInstaller,
+    FetchedScript,
+    CssSideEffectCallbackArgument,
+    CssSideEffectCallback,
+    ScriptSideEffectCallback,
+    ScriptSideEffectCallbackArgument,
+} from './inputs.models'
+export * from './errors.models'
+export * from './events.models'
+export { getAssetId, getUrlBase, parseResourceId } from './utils'
 export { youwolSvgLogo } from './utils.view'
 export * from './add-ons'
