+++ src/lib/inputs.models.ts
@@ -12,6 +12,8 @@
  * E.g.: `codemirror#5.52.0~mode/javascript.min.js`
  *
  */
+import { CdnEvent, CdnFetchEvent } from './events.models'
+
 export type FileLocationString = string
 
 /**
@@ -207,7 +209,6 @@ export type CustomInstaller = {
  *     document.getElementById('iFrameExampleModules').setAttribute("src",url);
  * </script>
  *
- * @category Entry Points
  */
 export type InstallInputs = {
     /**
@@ -310,7 +311,6 @@ export type InstallInputs = {
 /**
  * Inputs for the method {@link Client.fetchScript}
  *
- * @category Entry Points
  */
 export type FetchScriptInputs = {
     /**
@@ -334,7 +334,6 @@ export type FetchScriptInputs = {
 /**
  * Inputs for the method [[Client.installModules]]
  *
- * @category Entry Points
  */
 export type InstallModulesInputs = {
     /**
@@ -392,8 +391,6 @@ export type InstallPythonModulesInputs = {
 
 /**
  * Inputs for the method [[Client.installScripts]]
- *
- * @category Entry Points
  */
 export type InstallScriptsInputs = {
     /**
@@ -519,404 +516,6 @@ export type QueryLoadingGraphInputs = {
      * See [[InstallInputs.usingDependencies]]
      */
     usingDependencies?: LightLibraryQueryString[]
-}
-
-/**
- * Base class of Errors.
- *
- * @category Errors
- */
-export class CdnError extends Error {}
-
-/**
- * Base class of errors related to loading graph resolution. See also [[CdnLoadingGraphErrorEvent]].
- *
- * @category Errors
- */
-export class LoadingGraphError extends CdnError {
-    constructor() {
-        super('Failed to retrieve the loading graph') // (1)
-        this.name = 'LoadingGraphError' // (2)
-    }
-}
-
-/**
- * Error related to 401 response. See also [[UnauthorizedEvent]].
- *
- * @category Errors
- */
-export class Unauthorized extends CdnError {
-    static exceptionType = 'Unauthorized'
-
-    constructor(public readonly detail: { assetId; name; url }) {
-        super()
-    }
-
-    static isInstance(resp): resp is Unauthorized {
-        return resp.exceptionType == Unauthorized.exceptionType
-    }
-}
-
-/**
- * Error related to 404 response. See also [[UrlNotFoundEvent]].
- *
- * @category Errors
- */
-export class UrlNotFound extends CdnError {
-    static exceptionType = 'UrlNotFound'
-
-    constructor(public readonly detail: { assetId; name; url }) {
-        super()
-    }
-
-    static isInstance(resp): resp is UrlNotFound {
-        return resp.exceptionType == UrlNotFound.exceptionType
-    }
-}
-
-/**
- * Error happening while fetching a source file. See also [[FetchErrorEvent]].
- *
- * @category Errors
- */
-export class FetchErrors extends CdnError {
-    static exceptionType = 'FetchErrors'
-
-    constructor(public readonly detail: { errors }) {
-        super()
-    }
-
-    static isInstance(resp): resp is FetchErrors {
-        return resp.exceptionType == FetchErrors.exceptionType
-    }
-}
-
-/**
- * Error occurring while parsing a source content of a script. See also [[ParseErrorEvent]].
- *
- * @category Errors
- */
-export class SourceParsingFailed extends CdnError {
-    static exceptionType = 'SourceParsingFailed'
-
-    constructor(public readonly detail: { assetId; name; url; message }) {
-        super()
-    }
-
-    static isInstance(resp): resp is SourceParsingFailed {
-        return resp.exceptionType == SourceParsingFailed.exceptionType
-    }
-}
-
-/**
- * Error occurred trying to resolve a direct or indirect dependency while resolving a loading graph.
- * See also [[CdnLoadingGraphErrorEvent]].
- *
- * @category Errors
- */
-export class DependenciesError extends LoadingGraphError {
-    static exceptionType = 'DependenciesError'
-
-    constructor(
-        public readonly detail: {
-            context: string
-            errors: {
-                query: string
-                fromPackage: { name: string; version: string }
-                detail: string
-            }[]
-        },
-    ) {
-        super()
-    }
-
-    static isInstance(resp): resp is DependenciesError {
-        return resp.exceptionType == DependenciesError.exceptionType
-    }
-}
-
-/**
- * Dependencies resolution while resolving a loading graph lead to a circular dependency problem.
- * See also [[CdnLoadingGraphErrorEvent]].
- *
- * @category Errors
- */
-export class CircularDependencies extends LoadingGraphError {
-    static exceptionType = 'CircularDependencies'
-
-    constructor(
-        public readonly detail: {
-            context: string
-            packages: { [key: string]: { name: string; version: string }[] }
-        },
-    ) {
-        super()
-    }
-
-    static isInstance(resp): resp is CircularDependencies {
-        return resp.exceptionType == CircularDependencies.exceptionType
-    }
-}
-
-/**
- * Errors factory.
- *
- * @category Errors
- */
-export function errorFactory(error) {
-    if (CircularDependencies.isInstance(error)) {
-        return new CircularDependencies(error.detail)
-    }
-    if (DependenciesError.isInstance(error)) {
-        return new DependenciesError(error.detail)
-    }
-    if (Unauthorized.isInstance(error)) {
-        return new Unauthorized(error.detail)
-    }
-    if (error.exceptionType === 'UpstreamResponseException') {
-        return errorFactory(error.detail)
-    }
-}
-/**
- * @category Events
- */
-export type EventType =
-    | 'CdnMessageEvent'
-    | 'StartEvent'
-    | 'SourceLoadingEvent'
-    | 'SourceLoadedEvent'
-    | 'SourceParsedEvent'
-    | 'InstallDoneEvent'
-    | 'UnauthorizedEvent'
-    | 'UrlNotFoundEvent'
-    | 'ParseErrorEvent'
-    | 'CdnLoadingGraphErrorEvent'
-
-/**
- * @category Events
- */
-export type EventStatus = 'Pending' | 'Succeeded' | 'Failed' | 'None'
-/**
- * Base class of events.
- *
- * @category Events
- */
-export type CdnEvent = {
-    step: EventType
-    id: string
-    text: string
-    status: EventStatus
-}
-
-/**
- * @category Events
- */
-export function isCdnEvent(event: unknown): event is CdnEvent {
-    const types: EventType[] = [
-        'CdnMessageEvent',
-        'StartEvent',
-        'SourceLoadingEvent',
-        'SourceLoadedEvent',
-        'SourceParsedEvent',
-        'InstallDoneEvent',
-        'UnauthorizedEvent',
-        'UrlNotFoundEvent',
-        'ParseErrorEvent',
-        'CdnLoadingGraphErrorEvent',
-    ]
-    return types.includes((event as CdnEvent).step)
-}
-
-/**
- * Generic custom CDN event.
- *
- * @category Events
- */
-export class CdnMessageEvent implements CdnEvent {
-    public readonly step = 'CdnMessageEvent'
-    constructor(
-        public readonly id: string,
-        public readonly text: string,
-        public readonly status: EventStatus = 'None',
-    ) {}
-}
-
-/**
- * Base class for CDN's HTTP request event
- *
- * @category Events
- */
-export type CdnFetchEvent = CdnEvent & {
-    id: string
-    assetId: string
-    url: string
-}
-
-/**
- * Event emitted when starting to fetch a script.
- *
- * @category Events
- */
-export class StartEvent implements CdnFetchEvent {
-    public readonly step = 'StartEvent'
-    public readonly id: string
-    public readonly text: string
-    public readonly status = 'Pending'
-    constructor(
-        public readonly targetName: string,
-        public readonly assetId: string,
-        public readonly url: string,
-    ) {
-        this.id = targetName
-        this.text = `${targetName}: start importing`
-    }
-}
-
-/**
- * Event emitted when a script's content is transferring over HTTP network.
- *
- * @category Events
- */
-export class SourceLoadingEvent implements CdnFetchEvent {
-    public readonly step = 'SourceLoadingEvent'
-    public readonly id: string
-    public readonly text: string
-    public readonly status = 'Pending'
-    constructor(
-        public readonly targetName: string,
-        public readonly assetId: string,
-        public readonly url: string,
-        public readonly progress: ProgressEvent<XMLHttpRequestEventTarget>,
-    ) {
-        this.id = targetName
-        this.text = `${targetName}: fetching over HTTP`
-    }
-}
-
-/**
- * Event emitted when a script's content transfer over HTTP network has completed.
- *
- * @category Events
- */
-export class SourceLoadedEvent implements CdnFetchEvent {
-    public readonly step = 'SourceLoadedEvent'
-    public readonly id: string
-    public readonly text: string
-    public readonly status = 'Pending'
-    constructor(
-        public readonly targetName: string,
-        public readonly assetId: string,
-        public readonly url: string,
-        public readonly progress: ProgressEvent<XMLHttpRequestEventTarget>,
-    ) {
-        this.id = targetName
-        this.text = `${targetName}: source fetched`
-    }
-}
-
-/**
- * Event emitted when a script's content has been parsed (installed).
- *
- * @category Events
- */
-export class SourceParsedEvent implements CdnFetchEvent {
-    public readonly step = 'SourceParsedEvent'
-    public readonly id: string
-    public readonly text: string
-    public readonly status = 'Succeeded'
-    constructor(
-        public readonly targetName: string,
-        public readonly assetId: string,
-        public readonly url: string,
-    ) {
-        this.id = targetName
-        this.text = `${targetName}: module/script imported`
-    }
-}
-
-/**
- * Event emitted when an [[Unauthorized]] error occurred.
- *
- * @category Events
- */
-export class UnauthorizedEvent implements CdnFetchEvent {
-    public readonly step = 'UnauthorizedEvent'
-    public readonly id: string
-    public readonly text: string
-    public readonly status = 'Failed'
-    constructor(
-        public readonly targetName: string,
-        public readonly assetId: string,
-        public readonly url: string,
-    ) {
-        this.id = targetName
-        this.text = `${targetName}: unauthorized to access the resource`
-    }
-}
-
-/**
- * Event emitted when an [[UrlNotFound]] error occurred.
- *
- * @category Events
- */
-export class UrlNotFoundEvent implements CdnFetchEvent {
-    public readonly step = 'UrlNotFoundEvent'
-    public readonly id: string
-    public readonly text: string
-    public readonly status = 'Failed'
-    constructor(
-        public readonly targetName: string,
-        public readonly assetId: string,
-        public readonly url: string,
-    ) {
-        this.id = targetName
-        this.text = `${targetName}: resource not found at ${url}`
-    }
-}
-
-/**
- * Event emitted when an [[SourceParsingFailed]] error occurred.
- *
- * @category Events
- */
-export class ParseErrorEvent implements CdnFetchEvent {
-    public readonly step = 'UrlNotFoundEvent'
-    public readonly id: string
-    public readonly text: string
-    public readonly status = 'Failed'
-    constructor(
-        public readonly targetName: string,
-        public readonly assetId: string,
-        public readonly url: string,
-    ) {
-        this.id = targetName
-        this.text = `${targetName}: parsing the module/script failed`
-    }
-}
-
-/**
- * Event emitted when an [[LoadingGraphError]] error occurred.
- *
- * @category Events
- */
-export class CdnLoadingGraphErrorEvent implements CdnEvent {
-    public readonly id = 'loading-graph'
-    public readonly step = 'CdnLoadingGraphErrorEvent'
-    public readonly text = 'Failed to retrieve the loading graph'
-    public readonly status = 'Failed'
-    constructor(public readonly error: LoadingGraphError) {}
-}
-
-/**
- * Event emitted when an installation is done, see [[install]] & [[Client.install]].
- *
- * @category Events
- */
-export class InstallDoneEvent implements CdnEvent {
-    public readonly id = 'InstallDoneEvent'
-    public readonly step = 'InstallDoneEvent'
-    public readonly text = 'Installation successful'
-    public readonly status = 'Succeeded'
 }
 
 export type Library = {
